+++ R/grouped_ggscatterstats.R
@@ -52,6 +52,10 @@
 #'   the resolution of the data.
 #' @param axes.range.restrict Logical decides whether to restrict the axes values
 #'   ranges to min and max values of the `x` and `y` variables (Default: `FALSE`).
+#' @param ggtheme A function, `ggplot2` theme name. Default value is
+#'   `ggplot2::theme_grey()`. Allowed values are the official `ggplot2` themes,
+#'   including `theme_bw()`, `theme_minimal()`, `theme_classic()`,
+#'   `theme_void()`, etc.
 #' @param messages Decides whether messages references, notes, and warnings are
 #'   to be displayed (Default: `TRUE`).
 #' @inheritDotParams combine_plots
@@ -123,6 +127,7 @@ grouped_ggscatterstats <- function(grouping.var,
                                    maxit = 500,
                                    k = 3,
                                    axes.range.restrict = FALSE,
+                                   ggtheme = ggplot2::theme_grey(),
                                    messages = TRUE,
                                    ...) {
   # ========================================= preparing dataframe =======================================================
@@ -187,6 +192,7 @@ grouped_ggscatterstats <- function(grouping.var,
             maxit = maxit,
             k = k,
             axes.range.restrict = axes.range.restrict,
+            ggtheme = ggtheme,
             messages = messages
           )
         )